+++ spec/features/visitors/anyone_can_see_the_map_of_journey_steps_page_spec.rb
@@ -27,4 +27,30 @@ feature "Users can see all the steps of a journey" do
       end
     end
   end
+
+  context "when the map isn't valid" do
+    context "when the same entry is found twice" do
+      around do |example|
+        ClimateControl.modify(
+          CONTENTFUL_PLANNING_START_ENTRY_ID: "5kZ9hIFDvNCEhjWs72SFwj"
+        ) do
+          example.run
+        end
+      end
+
+      it "returns an error message" do
+        stub_get_contentful_entries(
+          entry_id: "5kZ9hIFDvNCEhjWs72SFwj",
+          fixture_filename: "repeat-entry-example.json"
+        )
+
+        visit new_journey_map_path
+
+        expect(page).to have_content(I18n.t("errors.repeat_step_in_the_contentful_journey.page_title"))
+        expect(page).to have_content(
+          I18n.t("errors.repeat_step_in_the_contentful_journey.page_body", entry_id: "5kZ9hIFDvNCEhjWs72SFwj")
+        )
+      end
+    end
+  end
 end